moved pysnobal and smrf_connect to awsm.models
