--- awsm/data/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-__author__ = """Micah Sandusky"""
-__email__ = 'micah.sandusky@ars.usda.gov'
-
-from . import init_model
