+++ awsm/models/pysnobal/init_model.py
