+++ awsm/tests/basins/RME/config.ini
@@ -245,7 +245,7 @@ folder_date_style:             start_end
 
 [awsm system]
 log_level:                     debug
-log_to_file:                   False
+log_to_file:                   True
 run_for_nsteps:                None
 output_frequency:              1
 daily_folders:                 False
